+++ src/Entity/Tile.php
@@ -7,6 +7,7 @@ use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\GetCollection;
 use App\Repository\TileRepository;
 use Doctrine\ORM\Mapping as ORM;
+use Gedmo\Timestampable\Traits\TimestampableEntity;
 
 #[ORM\Entity(repositoryClass: TileRepository::class)]
 #[ApiResource(
@@ -18,6 +19,8 @@ use Doctrine\ORM\Mapping as ORM;
 )]
 class Tile
 {
+    use TimestampableEntity;
+
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
@@ -38,8 +41,8 @@ class Tile
     #[ORM\Column(length: 255)]
     private ?string $mail = null;
 
-    #[ORM\Column(nullable: true)]
-    private array $extra = [];
+    #[ORM\Column(type: 'json', nullable: true)]
+    private ?string $extra = '';
 
     #[ORM\Column]
     private ?bool $accepted = null;
@@ -109,12 +112,12 @@ class Tile
         return $this;
     }
 
-    public function getExtra(): array
+    public function getExtra(): ?string
     {
         return $this->extra;
     }
 
-    public function setExtra(?array $extra): self
+    public function setExtra(?string $extra): self
     {
         $this->extra = $extra;